backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
starlette
